Tests, user session expiry time, small fixes

+++ src/fileserver/fileserver.go
@@ -13,13 +13,16 @@ import (
 	"os"
 	"strconv"
 	"strings"
+	"time"
 )
 
 const (
 	API_PREFIX        = "/api/"
+	GET_PREFIX        = "get/"
 	POST_PREFIX       = "post/"
 	DELETE_PREFIX     = "delete/"
 	DOWNLOAD_PREFIX   = "download/"
+	HOMEPAGE_PREFIX   = "home/"
 	UPLOAD_PREFIX     = "upload/"
 	CREATE_DIR_PREFIX = "createdir/"
 	NAVIGATION_PREFIX = "navigation/"
@@ -31,12 +34,14 @@ const (
 )
 
 var (
-	ROOT_DIR        string = "root"
-	NAVIGATION_FWD         = "fwd"
-	NAVIGATION_BACK        = "back"
-	USER_LOGIN_DATA        = make(map[string]bool)
-	USERS           []user
-	USER_DIRS       = make(map[string]string)
+	ROOT_DIR         string = "root"
+	SESSION_TIME     int64  = 0
+	NAVIGATION_FWD          = "fwd"
+	NAVIGATION_BACK         = "back"
+	USER_LOGIN_DATA         = make(map[string]*loginData)
+	ACTIVE_ADDRESSES        = make(map[string]string)
+	USERS            []user
+	USER_DIRS        = make(map[string]string)
 )
 
 /* --- types --- */
@@ -61,13 +66,19 @@ type user struct {
 	Password string
 }
 
+type loginData struct {
+	LoggedIn  bool
+	LoginTime time.Time
+}
+
 type configFile struct {
 	Config configuration `json:"config"`
 	Users  []user        `json:"users"`
 }
 
 type configuration struct {
-	RootDir string `json:"rootDir"`
+	RootDir     string `json:"rootDir"`
+	SessionTime int64  `json:"sessionTime"`
 }
 
 type fileInfo struct {
@@ -85,22 +96,44 @@ type dirInfo struct {
 }
 
 /* --- HTTP handlers --- */
-func downloadHandler(w http.ResponseWriter, r *http.Request) {
-	parameters := strings.Split(strings.Replace(r.URL.String(), API_PREFIX+DOWNLOAD_PREFIX, "", 1), "&")
-	username := strings.Split(parameters[0], "=")[1]
-	fileName := parameters[1]
-	if hasAccess(username) {
-		filePath := USER_DIRS[username] + string(os.PathSeparator) + fileName
-		serveFileForDownload(w, r, filePath)
-	}
-	return
-}
 
-func uploadHandler(w http.ResponseWriter, r *http.Request) {
-	username := strings.Replace(r.URL.String(), API_PREFIX+POST_PREFIX+UPLOAD_PREFIX, "", 1)
-	if !hasAccess(username) {
+/*
+	Handler for GET requests on /api/get/download/
+*/
+func downloadHandler(w http.ResponseWriter, r *http.Request) {
+	params := strings.Split(strings.Replace(r.URL.String(), API_PREFIX+GET_PREFIX+DOWNLOAD_PREFIX, "", 1), "&")
+	if len(params) < 2 {
+		fmt.Fprintln(w, "Invalid URL parameters!")
 		return
 	}
+	username := params[0]
+	fileName := params[1]
+
+	if acc, msg := hasAccess(username, r.RemoteAddr); !acc {
+		fmt.Fprintln(w, msg)
+		return
+	}
+
+	filePath := USER_DIRS[username] + string(os.PathSeparator) + fileName
+	serveFileForDownload(w, r, filePath)
+}
+
+/*
+	Handler for POST requests on /api/post/upload/
+*/
+func uploadHandler(w http.ResponseWriter, r *http.Request) {
+	username := strings.Replace(r.URL.String(), API_PREFIX+POST_PREFIX+UPLOAD_PREFIX, "", 1)
+
+	if len(username) == 0 {
+		fmt.Fprintln(w, "Invalid URL parameters!")
+		return
+	}
+
+	if acc, msg := hasAccess(username, r.RemoteAddr); !acc {
+		fmt.Fprintln(w, msg)
+		return
+	}
+
 	file, header, err := r.FormFile("file")
 	if err != nil {
 		fmt.Println(err.Error())
@@ -121,14 +154,24 @@ func uploadHandler(w http.ResponseWriter, r *http.Request) {
 			}
 		}
 	}
-	return
 }
 
+/*
+	Handler for POST requests on /api/post/createdir/
+*/
 func createDirHandler(w http.ResponseWriter, r *http.Request) {
 	username := strings.Replace(r.URL.String(), API_PREFIX+POST_PREFIX+CREATE_DIR_PREFIX, "", 1)
-	if !hasAccess(username) {
+
+	if len(username) == 0 {
+		fmt.Fprintln(w, "Invalid URL parameters!")
 		return
 	}
+
+	if acc, msg := hasAccess(username, r.RemoteAddr); !acc {
+		fmt.Fprintln(w, msg)
+		return
+	}
+
 	dirNameBytes, _ := ioutil.ReadAll(r.Body)
 	dirName := string(dirNameBytes)
 
@@ -146,18 +189,23 @@ func createDirHandler(w http.ResponseWriter, r *http.Request) {
 			fmt.Fprintf(w, jsonDirInfo)
 		}
 	}
-
-	return
 }
 
+/*
+	Handler for POST requests on /api/post/navigation/
+*/
 func navigationHandler(w http.ResponseWriter, r *http.Request) {
-	// TODO: check params for consistency
 	params := strings.Split(strings.Replace(r.URL.String(), API_PREFIX+POST_PREFIX+NAVIGATION_PREFIX, "", 1), "&")
+	if len(params) < 3 {
+		fmt.Fprintln(w, "Invalid URL parameters!")
+		return
+	}
+
 	direction := params[0]
 	username := params[1]
 
-	if !hasAccess(username) {
-		// TODO: err msg
+	if acc, msg := hasAccess(username, r.RemoteAddr); !acc {
+		fmt.Fprintln(w, msg)
 		return
 	}
 
@@ -178,83 +226,95 @@ func navigationHandler(w http.ResponseWriter, r *http.Request) {
 			fmt.Fprintf(w, dirinfo)
 		}
 	} else {
-		// something's wrong
+		fmt.Fprintln(w, "Invalid navigation!")
 	}
 }
 
-func navigateBack(username string) (string, error) {
-	currentDirPath := USER_DIRS[username]
-	pathEntries := strings.Split(currentDirPath, string(os.PathSeparator))
-	newDirPath := strings.Replace(currentDirPath, string(os.PathSeparator)+pathEntries[len(pathEntries)-1], "", 1)
+/*
+	Handler for GET requests on /api/get/home/
+*/
+func loadHomePageHandler(w http.ResponseWriter, r *http.Request) {
+	username := ACTIVE_ADDRESSES[r.RemoteAddr]
 
-	jsonDirInfo, e := generateDirectoryInfo(newDirPath)
-
-	return jsonDirInfo, e
-}
-
-func openDir(username string, dirname string) (string, error) {
-	dirPath := ROOT_DIR + string(os.PathSeparator) + USER_DIRECTORY_PREFIX + username + string(os.PathSeparator) + dirname
-	USER_DIRS[username] = dirPath
+	dirPath := USER_DIRS[username]
 	jsonDirInfo, e := generateDirectoryInfo(dirPath)
 
-	return jsonDirInfo, e
+	if e == nil {
+		fmt.Fprintf(w, username+"&"+jsonDirInfo)
+	}
 }
 
+/*
+	Handler for POST requests on /api/post/login/
+*/
 func loginHandler(w http.ResponseWriter, r *http.Request) {
-	fmt.Println(r.RemoteAddr)
 	body, _ := ioutil.ReadAll(r.Body)
 	var u user
 
 	err := json.Unmarshal(body, &u)
 	if err != nil {
-		fmt.Println(err.Error())
+		log.Println(err.Error())
+		return
 	}
 
-	if !loginUser(u) {
-		fmt.Fprintf(w, "failed&Invalid username or password")
+	if success, msg := loginUser(&u, r.RemoteAddr); !success {
+		fmt.Fprintf(w, "failed&"+msg)
 	} else {
 		dirName := ROOT_DIR + string(os.PathSeparator) + USER_DIRECTORY_PREFIX + u.Username
 		jsonDirInfo, e := generateDirectoryInfo(dirName)
 
 		if e != nil {
-			fmt.Fprintf(w, e.Error())
+			log.Println(e.Error())
+			fmt.Fprintf(w, "Cannot find user directory. Perhaps the root server directory has been moved?")
 		} else {
 			fmt.Fprintf(w, "success&"+jsonDirInfo)
 		}
 	}
-
-	return
 }
 
+/*
+	Handler for POST requests on /api/post/signup/
+*/
 func signupHandler(w http.ResponseWriter, r *http.Request) {
-	// TODO: extract function
 	body, _ := ioutil.ReadAll(r.Body)
 	var newUser user
 
 	err := json.Unmarshal(body, &newUser)
 	if err != nil {
-		fmt.Println(err.Error())
+		log.Println(err.Error())
+		return
 	}
 
 	err = signupUser(newUser)
 	if err != nil {
 		fmt.Fprintf(w, err.Error())
+		return
 	}
-	os.Mkdir(ROOT_DIR+string(os.PathSeparator)+USER_DIRECTORY_PREFIX+newUser.Username, 0777)
 
-	return
+	err = createUserDirectory(newUser.Username)
+	if err != nil {
+		fmt.Fprintf(w, err.Error())
+	}
 }
 
+/*
+	Handler for DELETE requests on /api/delete/
+*/
 func delHandler(w http.ResponseWriter, r *http.Request) {
 	params := strings.Split(strings.Replace(r.URL.String(), API_PREFIX+DELETE_PREFIX, "", 1), "&")
 	username := params[0]
 	itemName := params[1]
 
+	if acc, msg := hasAccess(username, r.RemoteAddr); !acc {
+		fmt.Fprintln(w, msg)
+		return
+	}
+
 	path := ROOT_DIR + string(os.PathSeparator) + USER_DIRECTORY_PREFIX + username + string(os.PathSeparator) + itemName
 	err := os.RemoveAll(path)
 
 	if err != nil {
-		fmt.Println(err.Error())
+		log.Println(err.Error())
 		fmt.Fprintf(w, err.Error())
 	} else {
 		dirName := ROOT_DIR + string(os.PathSeparator) + USER_DIRECTORY_PREFIX + username
@@ -266,15 +326,45 @@ func delHandler(w http.ResponseWriter, r *http.Request) {
 			fmt.Fprintf(w, jsonDirInfo)
 		}
 	}
-
-	return
 }
 
+/*
+	Retrieves folder data for the parent directory of the provided user's currently active one
+*/
+func navigateBack(username string) (string, error) {
+	currentDirPath := USER_DIRS[username]
+	if len(currentDirPath) == 0 {
+		return "", new(userNotFoundError)
+	}
+	pathEntries := strings.Split(currentDirPath, string(os.PathSeparator))
+	newDirPath := strings.Replace(currentDirPath, string(os.PathSeparator)+pathEntries[len(pathEntries)-1], "", 1)
+
+	jsonDirInfo, e := generateDirectoryInfo(newDirPath)
+
+	return jsonDirInfo, e
+}
+
+/*
+	Retrieves folder data for the child directory with the provided name (dirname)
+*/
+func openDir(username string, dirname string) (string, error) {
+	dirPath := ROOT_DIR + string(os.PathSeparator) + USER_DIRECTORY_PREFIX + username + string(os.PathSeparator) + dirname
+	jsonDirInfo, e := generateDirectoryInfo(dirPath)
+	if e == nil {
+		USER_DIRS[username] = dirPath
+	}
+
+	return jsonDirInfo, e
+}
+
+/*
+	Adds the new file to the filesystem.
+*/
 func uploadFileToServer(file multipart.File, header *multipart.FileHeader, username string) (err error) {
 	dir := USER_DIRS[username]
 	out, err := os.Create(dir + string(os.PathSeparator) + header.Filename)
 	if err != nil {
-		fmt.Println(err.Error())
+		log.Println(err.Error())
 		return err
 	}
 
@@ -282,18 +372,33 @@ func uploadFileToServer(file multipart.File, header *multipart.FileHeader, usern
 
 	_, err = io.Copy(out, file)
 	if err != nil {
-		fmt.Println(err.Error())
+		log.Println(err.Error())
 		return err
 	}
 
 	return nil
 }
 
+/*
+	A horse walks into a bar..
+*/
 func serveFileForDownload(w http.ResponseWriter, r *http.Request, filepath string) {
 	http.ServeFile(w, r, filepath)
 	return
 }
 
+/*
+	Creates a directory on the filesystem for the provided user in the following format: user-<username>
+*/
+func createUserDirectory(username string) error {
+	err := os.Mkdir(ROOT_DIR+string(os.PathSeparator)+USER_DIRECTORY_PREFIX+username, 0777)
+	return err
+}
+
+/*
+	Generates a string representation(json) of the file structure in the specified directory.
+	The result from this method contains file/directory names and file sizes.
+*/
 func generateDirectoryInfo(dirName string) (string, error) {
 	dirInfo, e := readDirectory(dirName)
 	if e != nil {
@@ -310,6 +415,10 @@ func generateDirectoryInfo(dirName string) (string, error) {
 	}
 }
 
+/*
+	Returns a pointer to a dirInfo object containing the data for the specified
+	folder (file/directory names and file sizes)
+*/
 func readDirectory(dirName string) (*dirInfo, error) {
 	items, err := ioutil.ReadDir(dirName)
 	var directoryInfo *dirInfo
@@ -339,26 +448,84 @@ func readDirectory(dirName string) (*dirInfo, error) {
 	return directoryInfo, nil
 }
 
-func hasAccess(username string) bool {
-	return USER_LOGIN_DATA[username]
+/*
+	Validates whether the provided user has access to the server.
+
+	The user can only access the server from the IP addresses from which
+	a login occured.
+	If the session time (see config file) for the provided user has expired
+	automatic logout gets triggered.
+*/
+func hasAccess(username string, addr string) (bool, string) {
+	if username == "" || addr == "" {
+		return false, "Access denied!"
+	}
+
+	loginData := USER_LOGIN_DATA[username]
+
+	if loginData != nil && loginData.LoggedIn {
+		if int64(time.Since(loginData.LoginTime).Seconds()) > SESSION_TIME {
+			user, _ := getUser(username)
+			logoutUser(&user)
+			return false, "Session timed out!"
+		}
+
+		if ACTIVE_ADDRESSES[addr] == username {
+			return true, ""
+		}
+	}
+
+	return false, "Access denied!"
 }
 
-func loginUser(u user) bool {
+/*
+	Checks whether the provided credentials are valid (present in the list of
+	registered users, see config file). If so, performs a login,
+	which consists of setting a timestamp and recording the IP address.
+	Also sets the active directory for the provided user to its root value (user-<username>)
+*/
+func loginUser(u *user, addr string) (bool, string) {
 	user, err := getUser(u.Username)
 	if err != nil {
-		fmt.Println(err.Error())
-		return false
+		log.Println(err.Error())
+		return false, err.Error()
 	}
 
 	if user.Password == u.Password {
-		USER_LOGIN_DATA[user.Username] = true
+		ld := USER_LOGIN_DATA[user.Username]
+		if ld == nil {
+			ld = new(loginData)
+			ld.LoggedIn = true
+			ld.LoginTime = time.Now()
+
+			USER_LOGIN_DATA[user.Username] = ld
+		}
+
+		ACTIVE_ADDRESSES[addr] = user.Username
+
 		USER_DIRS[user.Username] = ROOT_DIR + string(os.PathSeparator) + USER_DIRECTORY_PREFIX + user.Username
-		return true
+		return true, ""
 	}
 
-	return false
+	return false, "Invalid username or password"
 }
 
+/*
+	Clears all login data stored for the provided user
+*/
+func logoutUser(u *user) {
+	USER_LOGIN_DATA[u.Username] = nil
+	for key, _ := range ACTIVE_ADDRESSES {
+		if ACTIVE_ADDRESSES[key] == u.Username {
+			delete(ACTIVE_ADDRESSES, key)
+		}
+	}
+}
+
+/*
+	Retrieves the 'user' object for the provided username.
+	If the user is not present returns a 'userNotFoundError'
+*/
 func getUser(username string) (user, error) {
 	var u user
 	for _, user := range USERS {
@@ -372,6 +539,10 @@ func getUser(username string) (user, error) {
 	return u, err
 }
 
+/*
+	Performs a signup for the provided user. The operation updates the server
+	configuration file with the new user's data.
+*/
 func signupUser(newUser user) error {
 	_, e := getUser(newUser.Username)
 	if e == nil {
@@ -390,16 +561,18 @@ func signupUser(newUser user) error {
 	ioutil.WriteFile("config.json", jsonConfiguration, 0777)
 
 	USERS = newUsersArray
-	USER_LOGIN_DATA[newUser.Username] = false
 	return nil
 }
 
+/*
+	Reads the server configuration file and emits a 'configFile' object.
+*/
 func loadConfiguration() configFile {
 	path, _ := os.Getwd()
 	file, err := os.Open(path + string(os.PathSeparator) + "config.json")
 	defer file.Close()
 	if err != nil {
-		fmt.Println(err.Error())
+		log.Println(err.Error())
 		log.Fatal("Failed to load configuration!")
 	}
 
@@ -410,34 +583,45 @@ func loadConfiguration() configFile {
 	// read config file
 	err = json.Unmarshal(buffer.Bytes(), &configFile)
 	if err != nil {
-		fmt.Println(err.Error())
+		log.Println(err.Error())
 		log.Fatal("Configuration file is corrupt!")
 	}
 	return configFile
 }
 
+/*
+	Sets server runtime parameters based on the provided configuration.
+*/
 func applyConfiguration(config configuration) {
 	ROOT_DIR = config.RootDir
-}
-
-func initLoginData(users []user) {
-	for _, user := range users {
-		USER_LOGIN_DATA[user.Username] = false
+	SESSION_TIME = config.SessionTime
+	if _, err := os.Stat(ROOT_DIR); os.IsNotExist(err) {
+		err = os.Mkdir(ROOT_DIR, 0777)
+		if err != nil {
+			log.Println(err.Error())
+		}
 	}
 }
 
+/*
+	Initializes the server with all HTTP handlers.
+*/
 func initServer() http.Handler {
 	server := pat.New()
-	server.Get(API_PREFIX+DOWNLOAD_PREFIX, http.HandlerFunc(downloadHandler))
 
-	secondaryHandler := pat.New()
-	secondaryHandler.Post(API_PREFIX+POST_PREFIX+UPLOAD_PREFIX, http.HandlerFunc(uploadHandler))
-	secondaryHandler.Post(API_PREFIX+POST_PREFIX+CREATE_DIR_PREFIX, http.HandlerFunc(createDirHandler))
-	secondaryHandler.Post(API_PREFIX+POST_PREFIX+NAVIGATION_PREFIX, http.HandlerFunc(navigationHandler))
-	secondaryHandler.Post(API_PREFIX+POST_PREFIX+LOGIN_PREFIX, http.HandlerFunc(loginHandler))
-	secondaryHandler.Post(API_PREFIX+POST_PREFIX+SIGNUP_PREFIX, http.HandlerFunc(signupHandler))
+	getServer := pat.New()
+	getServer.Get(API_PREFIX+GET_PREFIX+DOWNLOAD_PREFIX, http.HandlerFunc(downloadHandler))
+	getServer.Get(API_PREFIX+GET_PREFIX+HOMEPAGE_PREFIX, http.HandlerFunc(loadHomePageHandler))
 
-	server.Post(API_PREFIX+POST_PREFIX, secondaryHandler)
+	postServer := pat.New()
+	postServer.Post(API_PREFIX+POST_PREFIX+UPLOAD_PREFIX, http.HandlerFunc(uploadHandler))
+	postServer.Post(API_PREFIX+POST_PREFIX+CREATE_DIR_PREFIX, http.HandlerFunc(createDirHandler))
+	postServer.Post(API_PREFIX+POST_PREFIX+NAVIGATION_PREFIX, http.HandlerFunc(navigationHandler))
+	postServer.Post(API_PREFIX+POST_PREFIX+LOGIN_PREFIX, http.HandlerFunc(loginHandler))
+	postServer.Post(API_PREFIX+POST_PREFIX+SIGNUP_PREFIX, http.HandlerFunc(signupHandler))
+
+	server.Get(API_PREFIX+GET_PREFIX, getServer)
+	server.Post(API_PREFIX+POST_PREFIX, postServer)
 	server.Del(API_PREFIX+DELETE_PREFIX, http.HandlerFunc(delHandler))
 
 	return server
@@ -448,7 +632,6 @@ func main() {
 	configFile := loadConfiguration()
 	configuration := configFile.Config
 	USERS = configFile.Users
-	initLoginData(USERS)
 	applyConfiguration(configuration)
 
 	// host client
@@ -456,5 +639,5 @@ func main() {
 	// handle file requests
 	http.Handle(API_PREFIX, initServer())
 	//start listening
-	http.ListenAndServe(":"+strconv.Itoa(PORT), nil)
+	log.Fatal(http.ListenAndServe(":"+strconv.Itoa(PORT), nil))
 }
